Add duplicacy backup tool
I have not used this tool, but what is interesting about it is the lock-free deduplication (unique?) feature that it is claimed to support. It is also written in golang, so, perhaps it can give some ideas to restic authors if they have the time to look into its internals.